requirements.txt
attrs==22.2.0
boxsdk==3.7.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
idna==3.4
pur==7.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
requests-toolbelt==0.10.1
six==1.16.0
urllib3==1.26.15
